+++ php-plugin/src/main/java/fr/greencodeinitiative/php/PHPPlugin.java
@@ -1,7 +1,5 @@
 /*
- * SonarQube PHP Custom Rules Example
- * Copyright (C) 2016-2016 SonarSource SA
- * mailto:contact AT sonarsource DOT com
+ * Copyright (C) 2023 Green Code Initiative
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -13,21 +11,18 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
  *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 package fr.greencodeinitiative.php;
 
 import org.sonar.api.Plugin;
 
-/**
- * Extension point to define a Sonar Plugin.
- */
 public class PHPPlugin implements Plugin {
 
     @Override
     public void define(Context context) {
         context.addExtension(PhpRuleRepository.class);
     }
+
 }